OFMapTable: Fix type mismatch on OFHashSeed
FossilOrigin-Name: 0cc973c595ccd95f272a46e51c71f1d8f59a9ca097e9e515516451d5fb35d690

+++ src/OFMapTable.m
@@ -28,7 +28,7 @@
 #import "OFInvalidArgumentException.h"
 #import "OFOutOfRangeException.h"
 
-extern uint32_t OFHashSeed;
+extern unsigned long OFHashSeed;
 
 static const uint32_t minCapacity = 16;
